requirements.txt
# ============================================================================
# BCSS Medical Image Segmentation - Python Dependencies
# ============================================================================
# 
# 安裝指南：
# 1. 基礎: pip install -r requirements.txt
# 2. MMSegmentation: pip install mmsegmentation
#    或從源碼: cd mmsegmentation && pip install -e .
# 3. (可選) CUDA 支持: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# Python 版本: >= 3.8（推薦 3.10+）
# ============================================================================

# ============================================================================
# 深度學習核心框架
# ============================================================================
torch>=2.0.0           # PyTorch - 深度學習框架
torchvision>=0.15.0    # PyTorch 視覺工具包
numpy>=1.23.0          # 數值計算

# ============================================================================
# OpenMMLab 生態系統
# ============================================================================
openmim>=0.3.9         # OpenMMLab 安裝管理器
mmengine>=0.7.4        # MMEngine - OpenMMLab 基礎庫
mmsegmentation>=1.0.0  # MMSegmentation - 分割框架
# 注意：mmsegmentation 也可單獨安裝：
#   pip install mmsegmentation
# 或從本地源碼安裝：
#   cd mmsegmentation && pip install -e .

# ============================================================================
# 影像處理與資料工具
# ============================================================================
opencv-python-headless>=4.7.0  # OpenCV (headless - 用於伺服器環境)
Pillow>=9.5.0                   # PIL - 影像操作
scikit-image>=0.20.0            # 進階影像處理
imageio>=2.25.0                 # 影像讀寫

# ============================================================================
# 資料處理與分析
# ============================================================================
pandas>=1.5.0     # 資料框架與分析
scikit-learn>=1.2.0  # 機器學習工具（metrics 等）

# ============================================================================
# 可視化與製圖
# ============================================================================
matplotlib>=3.7.0  # 繪圖庫（訓練曲線、結果展示）
seaborn>=0.12.0    # 統計視覺化

# ============================================================================
# 工具與實用程式
# ============================================================================
tqdm>=4.65.0       # 進度條
pyyaml>=6.0        # YAML 配置檔案
Pillow>=9.5.0      # 影像處理
requests>=2.28.0   # HTTP 請求（資料下載）
joblib             # 平行處理用 (added by preprocessing)

# ============================================================================
# (可選) 資料下載工具
# ============================================================================
# kaggle>=1.5.12   # Kaggle API（若需下載 BCSS 資料集）
# 使用前需配置 Kaggle 認證：~/.kaggle/kaggle.json

# ============================================================================
# (可選) 開發與測試工具
# ============================================================================
# pytest>=7.2.0     # 單元測試
# black>=23.0.0     # 代碼格式化
# flake8>=6.0.0     # 代碼風格檢查
# 安裝：pip install pytest black flake8

# ============================================================================
# 環境檢查
# ============================================================================
# 安裝後驗證：
#   python -c "import torch; print(f'PyTorch: {torch.__version__}')"
#   python -c "import mmseg; print('MMSegmentation: OK')"
#   python -c "import mmengine; print('MMEngine: OK')"
